BTC_Project/requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.3.2
statsmodels==0.14.1
yfinance==0.2.54
pmdarima==2.0.4
arch==6.3.0
ta==0.10.2
joblib==1.4.2
tqdm==4.66.4
holidays==0.48
scipy==1.12.0
requests==2.31.0
pyyaml==6.0.1
Flask==3.0.3
